refactor: server-info repository
