+++ server/src/types.ts
@@ -12,6 +12,7 @@ import { MediaRepository } from 'src/repositories/media.repository';
 import { MemoryRepository } from 'src/repositories/memory.repository';
 import { MetadataRepository } from 'src/repositories/metadata.repository';
 import { OAuthRepository } from 'src/repositories/oauth.repository';
+import { ServerInfoRepository } from 'src/repositories/server-info.repository';
 import { MetricGroupRepository, TelemetryRepository } from 'src/repositories/telemetry.repository';
 import { TrashRepository } from 'src/repositories/trash.repository';
 import { VersionHistoryRepository } from 'src/repositories/version-history.repository';
@@ -53,6 +54,7 @@ export type IMemoryRepository = RepositoryInterface<MemoryRepository>;
 export type IMetadataRepository = RepositoryInterface<MetadataRepository>;
 export type IMetricGroupRepository = RepositoryInterface<MetricGroupRepository>;
 export type IOAuthRepository = RepositoryInterface<OAuthRepository>;
+export type IServerInfoRepository = RepositoryInterface<ServerInfoRepository>;
 export type ITelemetryRepository = RepositoryInterface<TelemetryRepository>;
 export type ITrashRepository = RepositoryInterface<TrashRepository>;
 export type IViewRepository = RepositoryInterface<ViewRepository>;
